requirements.txt
asgiref==3.11.1
Django==5.2.11
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.11.0
resend==2.26.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
# Production dependencies
gunicorn==21.2.0
whitenoise==6.7.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
